Move configuration  classes
- Create package for configuration classes and move them there
- Adjust class name to represent it's function
