+++ src/main/java/org/phoebus/channelfinder/configuration/HttpConnectorConfig.java
@@ -1,4 +1,4 @@
-package org.phoebus.channelfinder;
+package org.phoebus.channelfinder.configuration;
 
 import org.apache.catalina.connector.Connector;
 import org.springframework.beans.factory.annotation.Value;
