+++ src/main/java/org/phoebus/channelfinder/configuration/RequestLoggingFilterConfig.java
@@ -1,4 +1,4 @@
-package org.phoebus.channelfinder;
+package org.phoebus.channelfinder.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
