+++ src/main/java/org/phoebus/channelfinder/respository/ChannelRepository.java
@@ -50,8 +50,8 @@ import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 import javax.annotation.PreDestroy;
 import org.phoebus.channelfinder.CFResourceDescriptors;
-import org.phoebus.channelfinder.ElasticConfig;
 import org.phoebus.channelfinder.TextUtil;
+import org.phoebus.channelfinder.configuration.ElasticConfig;
 import org.phoebus.channelfinder.entity.Channel;
 import org.phoebus.channelfinder.entity.Property;
 import org.phoebus.channelfinder.entity.SearchResult;
