+++ src/main/java/org/phoebus/channelfinder/respository/PropertyRepository.java
@@ -31,8 +31,8 @@ import java.util.Optional;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.StreamSupport;
-import org.phoebus.channelfinder.ElasticConfig;
 import org.phoebus.channelfinder.TextUtil;
+import org.phoebus.channelfinder.configuration.ElasticConfig;
 import org.phoebus.channelfinder.entity.Channel;
 import org.phoebus.channelfinder.entity.Property;
 import org.phoebus.channelfinder.entity.Property.OnlyNameOwnerProperty;
