+++ src/main/java/org/phoebus/channelfinder/respository/TagRepository.java
@@ -32,8 +32,8 @@ import java.util.Optional;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.StreamSupport;
-import org.phoebus.channelfinder.ElasticConfig;
 import org.phoebus.channelfinder.TextUtil;
+import org.phoebus.channelfinder.configuration.ElasticConfig;
 import org.phoebus.channelfinder.entity.Channel;
 import org.phoebus.channelfinder.entity.Tag;
 import org.phoebus.channelfinder.entity.Tag.OnlyTag;
