+++ src/test/java/org/phoebus/channelfinder/performance/ExistsPerformanceIT.java
@@ -3,7 +3,7 @@ package org.phoebus.channelfinder.performance;
 import com.google.common.collect.Lists;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.phoebus.channelfinder.example.PopulateService;
+import org.phoebus.channelfinder.configuration.PopulateDBConfiguration;
 import org.phoebus.channelfinder.respository.ChannelRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -16,7 +16,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 @WebMvcTest(ChannelRepository.class)
 class ExistsPerformanceIT {
 
-  @Autowired PopulateService service;
+  @Autowired PopulateDBConfiguration service;
 
   @Autowired ChannelRepository channelRepository;
