+++ src/test/java/org/phoebus/channelfinder/performance/PopulateDBConfigurationIT.java
@@ -2,7 +2,7 @@ package org.phoebus.channelfinder.performance;
 
 import java.net.URL;
 import org.junit.jupiter.api.Test;
-import org.phoebus.channelfinder.example.PopulateService;
+import org.phoebus.channelfinder.configuration.PopulateDBConfiguration;
 import org.phoebus.channelfinder.service.AuthorizationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -12,9 +12,9 @@ import org.springframework.test.context.TestPropertySource;
 /** An class for creating the example database and testing the speed. */
 @WebMvcTest(AuthorizationService.class)
 @TestPropertySource(value = "classpath:performance_application.properties")
-class PopulateServiceIT {
+class PopulateDBConfigurationIT {
 
-  @Autowired PopulateService populateService;
+  @Autowired PopulateDBConfiguration populateDBConfiguration;
 
   @Test
   @WithMockUser(username = "admin", roles = "CF-ADMINS")
@@ -22,16 +22,16 @@ class PopulateServiceIT {
     final URL tagResource = getClass().getResource("/perf_tags.json");
     final URL propertyResource = getClass().getResource("/perf_properties.json");
 
-    populateService.createTagsAndProperties(tagResource, propertyResource);
+    populateDBConfiguration.createTagsAndProperties(tagResource, propertyResource);
   }
 
   @Test
   void testCreateDB() {
-    populateService.createDB();
+    populateDBConfiguration.createDB();
   }
 
   @Test
   void testCleanUpDB() {
-    populateService.cleanupDB();
+    populateDBConfiguration.cleanupDB();
   }
 }
